+++ contracts/token/TokenAgeWeights.sol
@@ -8,12 +8,12 @@ import "../interfaces/IVotingWeights.sol";
 contract TokenAgeWeights is TokenAgeERC20, IVotingWeights {
     function weightOf(address owner) public view returns (uint256 weight) {
         Checkpoint storage checkpoint = _getLastCheckpoint(_checkpoints[owner]);
-        weight = uint256(_getTokenAge(checkpoint, _currentTime()));
+        weight = _getTokenAge(checkpoint, _currentTime());
     }
 
     function weightOfAt(address owner, uint256 atBlock) public override view returns (uint256 weight) {
         uint64 atTime = _convertTime(atBlock);
         Checkpoint storage checkpoint = _getCheckpoint(_checkpoints[owner], atTime);
-        weight = uint256(_getTokenAge(checkpoint, atTime));
+        weight = _getTokenAge(checkpoint, atTime);
     }
 }
